backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
databricks-sdk==0.32.0
databricks-sql-connector==3.4.0
python-multipart==0.0.9
